Extract HomepageNewsFeed component
